Leave 'albums_per_page' option for 3rd party themes

+++ npgCore/setup/setup_themeOptions.php
@@ -56,8 +56,11 @@ if (!empty($requirePath)) {
 }
 /* then set any "standard" options that may not have been covered by the theme */
 standardThemeOptions($theme, NULL);
-//	purge obsolete theme options
-purgeOption('albums_per_row', $theme);
+
+if (protectedTheme($theme)) {
+	//	purge obsolete theme options
+	purgeOption('albums_per_row', $theme);
+}
 
 sendImage($_GET['class'], 'theme_' . $theme);
